refactor(mito2): share parquet index writer
